requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
curlify==2.2.1
facebook-business==20.0.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pycountry==24.6.1
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
yarl==1.9.4
cryptography==42.0.8
Flask==2.3.3